python_backend/requirements.txt
Flask==2.3.3
opencv-python==4.7.0.72
deepface==0.0.75
speechrecognition==3.8.1
requests==2.28.1
python-dotenv==1.0.0